Centralize test setup in `TestSuite`
